requirements.txt
numpy>=1.21,<1.23
torch>=2.0
gym==0.19.0
numba>=0.55
pyglet<1.5
pyyaml
h5py
tqdm
matplotlib
